feat(blackjack) : DIVIDE - Dealer class
